+++ src/App.css
@@ -13,8 +13,3 @@
     background-color: #555;
     color: #ccc;
 }
-
-.App.dark input{
-    background-color: #222222;
-    color: #eeeeee;
-}
