+++ src/components/Amount.css
@@ -1,0 +1,4 @@
+.amount.dark input{
+    background-color: #555;
+    color: #ccc;  
+}
